backend/requirements.txt
fastapi
uvicorn
supabase
yfinance
pandas
pydantic
pydantic-settings
requests
python-dotenv
google-generativeai
pytest
httpx
